wip: barking up the wrong tree.

+++ src/pb/pbstore.c
@@ -32,16 +32,16 @@ void* pbstore_get()
 
 bool recents_exists() { return persist_exists(100000); }
 
-void recents_set(int* ids)
+void recents_set(short* ids)
 {
   persist_write_int(100000, (int) sizeof(*ids));
   persist_write_data(100001, ids, sizeof(*ids));
 };
 
-int* recents_get()
+short* recents_get()
 {
   size_t bytes = (size_t) persist_read_int(100000);
-  int* buffer = malloc(bytes);
+  short* buffer = malloc(bytes);
   persist_read_data(100001, buffer, bytes);
 };
